refactor: split files >1K LOC into impl + tests siblings
Hard ceiling of 1000 LOC per file. Mechanical split — for each src
file the inline `#[cfg(test)] mod tests` block moves to a sibling
`<name>_tests.rs`, re-declared via `#[cfg(test)] #[path = "..."] mod
tests;` so `super::*` still resolves to the parent module's private
items.

For tests/pty_behavior.rs the integration-test file is split into
tests/pty_behavior.rs (T1/T2/T3), tests/pty_pump.rs (raw-pump tests),
and tests/common/mod.rs (shared harness + require_pty_or_skip! macro).

Files affected (before -> after):
  src/session_registry.rs            1327 -> 768 + 559 tests
  src/session.rs                     1256 -> 937 + 319 tests
  src/dnd/console_drop_target.rs     1255 -> 999 + 256 tests
  src/hook_health.rs                 1172 -> 962 + 210 tests
  src/args.rs                        1165 -> 377 + 788 tests
  src/worktrees.rs                   1055 -> 668 + 387 tests
  src/gc.rs                          1028 -> 785 + 243 tests
  tests/pty_behavior.rs              1209 -> 380 + 633 pump + 229 common

All 567 Rust unit tests pass; cargo fmt / clippy -D warnings / ruff
all clean. No public API change.
