+++ crates/clud-bin/src/hook_health.rs
@@ -958,215 +958,5 @@ fn display_path(path: &Path) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::tempdir;
-
-    fn write(path: &Path, body: &str) {
-        fs::create_dir_all(path.parent().unwrap()).unwrap();
-        fs::write(path, body).unwrap();
-    }
-
-    fn trusted_state_for(path: &Path, group: usize, handler: usize) -> String {
-        format!(
-            "[hooks.state.'{}:{}:{group}:{handler}']\ntrusted_hash = \"sha256:test\"\n",
-            path.to_string_lossy(),
-            CODEX_PRE_TOOL_USE_STATE
-        )
-    }
-
-    #[test]
-    fn claude_settings_parser_normalizes_pre_tool_matchers() {
-        let temp = tempdir().unwrap();
-        let repo = temp.path().join("repo");
-        let home = temp.path().join("home");
-        write(
-            &repo.join(".claude").join("settings.json"),
-            r#"{"hooks":{"PreToolUse":[{"matcher":"Bash","hooks":[{"type":"command","command":"python check.py"}]}]}}"#,
-        );
-
-        let report = inspect_paths(&repo, Some(&home));
-
-        assert_eq!(
-            report.claude.all_matchers(),
-            BTreeSet::from(["Bash".to_string()])
-        );
-        assert!(report.claude.hooks[0].active);
-    }
-
-    #[test]
-    fn codex_current_shape_parses_and_legacy_shape_warns() {
-        let temp = tempdir().unwrap();
-        let repo = temp.path().join("repo");
-        let home = temp.path().join("home");
-        let current = repo.join(".codex").join("hooks.json");
-        write(
-            &current,
-            r#"{"hooks":{"PreToolUse":[{"matcher":"*","hooks":[{"type":"command","command":"python hook.py"}]}]}}"#,
-        );
-        write(
-            &home.join(".codex").join("config.toml"),
-            &format!(
-                "[projects.'{}']\ntrust_level = \"trusted\"\n{}",
-                codex_project_key(&repo),
-                trusted_state_for(&current, 0, 0)
-            ),
-        );
-
-        let report = inspect_paths(&repo, Some(&home));
-        assert_eq!(
-            report.codex.all_matchers(),
-            BTreeSet::from(["*".to_string()])
-        );
-        assert!(report.codex.hooks[0].active);
-
-        write(
-            &current,
-            r#"{"PreToolUse":[{"matcher":"Bash","hooks":[]}]}"#,
-        );
-        let report = inspect_paths(&repo, Some(&home));
-        assert!(report
-            .warnings
-            .iter()
-            .any(|warning| warning.contains("legacy root-level")));
-    }
-
-    #[test]
-    fn one_sided_hook_warnings_are_bidirectional() {
-        let temp = tempdir().unwrap();
-        let repo = temp.path().join("repo");
-        let home = temp.path().join("home");
-        write(
-            &repo.join(".claude").join("settings.json"),
-            r#"{"hooks":{"PreToolUse":[{"matcher":"Bash","hooks":[{}]}]}}"#,
-        );
-        let report = inspect_paths(&repo, Some(&home));
-        assert!(report
-            .warnings
-            .iter()
-            .any(|warning| warning.contains("Claude PreToolUse hooks exist")
-                && warning.contains("clud --fix-hooks")));
-
-        fs::remove_file(repo.join(".claude").join("settings.json")).unwrap();
-        let codex = repo.join(".codex").join("hooks.json");
-        write(
-            &codex,
-            r#"{"hooks":{"PreToolUse":[{"matcher":"Bash","hooks":[{}]}]}}"#,
-        );
-        write(
-            &home.join(".codex").join("config.toml"),
-            &format!(
-                "[projects.'{}']\ntrust_level = \"trusted\"\n{}",
-                codex_project_key(&repo),
-                trusted_state_for(&codex, 0, 0)
-            ),
-        );
-        let report = inspect_paths(&repo, Some(&home));
-        assert!(report
-            .warnings
-            .iter()
-            .any(|warning| warning.contains("Codex PreToolUse hooks exist")
-                && warning.contains("clud --fix-hooks")));
-    }
-
-    #[test]
-    fn matcher_mismatch_warns() {
-        let temp = tempdir().unwrap();
-        let repo = temp.path().join("repo");
-        let home = temp.path().join("home");
-        write(
-            &repo.join(".claude").join("settings.json"),
-            r#"{"hooks":{"PreToolUse":[{"matcher":"Bash","hooks":[{}]}]}}"#,
-        );
-        let codex = repo.join(".codex").join("hooks.json");
-        write(
-            &codex,
-            r#"{"hooks":{"PreToolUse":[{"matcher":"Read","hooks":[{}]}]}}"#,
-        );
-        write(
-            &home.join(".codex").join("config.toml"),
-            &format!(
-                "[projects.'{}']\ntrust_level = \"trusted\"\n{}",
-                codex_project_key(&repo),
-                trusted_state_for(&codex, 0, 0)
-            ),
-        );
-
-        let report = inspect_paths(&repo, Some(&home));
-
-        assert!(report
-            .warnings
-            .iter()
-            .any(|warning| warning.contains("matchers differ")));
-    }
-
-    #[test]
-    fn windows_codex_project_key_normalizes_extended_path() {
-        let path = PathBuf::from(r"\\?\C:\Users\Niteris\Dev\Repo");
-        assert_eq!(codex_project_key(&path), r"c:\users\niteris\dev\repo");
-        assert!(is_extended_key_for(
-            r"\\?\C:\Users\Niteris\Dev\Repo",
-            r"c:\users\niteris\dev\repo"
-        ));
-    }
-
-    #[test]
-    fn extended_path_only_trust_config_warns() {
-        let temp = tempdir().unwrap();
-        let repo = PathBuf::from(r"C:\Users\Niteris\Dev\Repo");
-        let home = temp.path().join("home");
-        let codex = repo.join(".codex").join("hooks.json");
-        write(
-            &codex,
-            r#"{"hooks":{"PreToolUse":[{"matcher":"*","hooks":[{}]}]}}"#,
-        );
-        write(
-            &home.join(".codex").join("config.toml"),
-            r#"[projects.'\\?\C:\Users\Niteris\Dev\Repo']
-trust_level = "trusted"
-"#,
-        );
-
-        let report = inspect_paths(&repo, Some(&home));
-
-        assert!(report
-            .warnings
-            .iter()
-            .any(|warning| warning.contains("extended `\\\\?\\...` path key")));
-    }
-
-    #[test]
-    fn deterministic_trust_repair_adds_canonical_key() {
-        let temp = tempdir().unwrap();
-        let config = temp.path().join("home").join(".codex").join("config.toml");
-        add_codex_project_trust(&config, r"c:\users\niteris\dev\repo").unwrap();
-
-        let text = fs::read_to_string(config).unwrap();
-        assert!(text.contains(r#"trust_level = "trusted""#));
-        assert!(text.contains(r#"c:\users\niteris\dev\repo"#));
-        assert!(!text.contains("trusted_hash"));
-    }
-
-    #[test]
-    fn per_hook_prompt_planning_uses_one_action_per_missing_matcher() {
-        let temp = tempdir().unwrap();
-        let repo = temp.path().join("repo");
-        let home = temp.path().join("home");
-        write(
-            &repo.join(".claude").join("settings.json"),
-            r#"{"hooks":{"PreToolUse":[{"matcher":"Bash","hooks":[{}]},{"matcher":"Read","hooks":[{}]}]}}"#,
-        );
-
-        let report = inspect_paths(&repo, Some(&home));
-        let prompts = plan_repairs(&report)
-            .into_iter()
-            .filter(|action| matches!(action, RepairAction::BackendPrompt { .. }))
-            .collect::<Vec<_>>();
-
-        assert_eq!(prompts.len(), 2);
-        assert!(prompts.iter().all(|action| match action {
-            RepairAction::BackendPrompt { prompt, .. } => prompt.contains("exactly one hook"),
-            _ => false,
-        }));
-    }
-}
+#[path = "hook_health_tests.rs"]
+mod tests;
